thorlab_powermeter_100USB/requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.1
PyVISA==1.14.1
scipy==1.13.1
six==1.16.0
ThorlabsPM100==1.2.2
typing_extensions==4.11.0
tzdata==2024.1
